# api/openapi.yaml
openapi: 3.1.0
info:
  title: TOT API
  version: 1.0.0-draft
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    JSON API for the Next.js frontend (`frontend-tot`). Draft contract — see
    `docs/frontend/inventory.md` for the page inventory and open decisions.

    Conventions:
    - Base path `/api/v1`; frontend and API share one origin (reverse proxy), so the
      session cookie stays `HttpOnly; SameSite=Lax` and no CORS is needed.
    - JSON bodies and fields are `snake_case`; timestamps are RFC 3339.
    - Errors always use the `Error` schema with a stable machine `code`; user-facing
      text is localized on the frontend by `code`, `message` is for logs only.
    - Unauthenticated requests get `401`, never a redirect; malformed JSON gets `400 malformed_json`.
    - Mutating requests (POST/PUT/PATCH/DELETE) must come from an allowed origin
      (`Origin` / `Sec-Fetch-Site`), otherwise `403 csrf_rejected`.
    - Request bodies are `application/json` unless stated otherwise, otherwise `415 unsupported_media_type`.
    - HTML fields (`*_html`) are rendered and sanitized server-side.
    - Admin endpoints need role `author` or `admin`; admin-only ones say so. A course
      the author cannot access answers `404`; a missing permission on an accessible
      course answers `403 forbidden` (see `CourseAccess`).
servers:
  - url: /api/v1
...
security:
  - session: []
paths:
...
  /admin/courses/{courseId}/lessons:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    post:
      tags: [admin-lessons]
      summary: Create lesson
      operationId: adminCreateLesson
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminLessonInput' }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminLesson' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/SlugTaken' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /admin/lessons/{lessonId}:
    parameters:
      - $ref: '#/components/parameters/LessonId'
    get:
      tags: [admin-lessons]
      summary: Get lesson
      operationId: adminGetLesson
      responses:
        '200':
          description: Lesson with its quiz questions and tasks (all fields).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminLessonDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [admin-lessons]
      summary: Update lesson
      operationId: adminUpdateLesson
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminLessonInput' }
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminLesson' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/SlugTaken' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    delete:
      tags: [admin-lessons]
      summary: Delete lesson
      operationId: adminDeleteLesson
      description: Cascades to quiz, tasks, submissions and progress. Published lessons need `can_publish`, drafts `can_edit`.
      responses:
        '204': { description: Deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/lessons/{lessonId}/published:
    parameters:
      - $ref: '#/components/parameters/LessonId'
    put:
      tags: [admin-lessons]
      summary: Set lesson published
      operationId: adminSetLessonPublished
      description: Needs `can_publish`.
      requestBody: { $ref: '#/components/requestBodies/Published' }
      responses:
        '204': { description: Updated. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/lessons/{lessonId}/move:
    parameters:
      - $ref: '#/components/parameters/LessonId'
    post:
      tags: [admin-lessons]
      summary: Move lesson
      operationId: adminMoveLesson
      requestBody: { $ref: '#/components/requestBodies/Move' }
      responses:
        '204': { description: Moved. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/lessons/{lessonId}/duplicate:
    parameters:
      - $ref: '#/components/parameters/LessonId'
    post:
      tags: [admin-lessons]
      summary: Duplicate lesson
      operationId: adminDuplicateLesson
      description: Copies lesson, quiz and tasks as a draft with slug `<slug>-copy[-N]`.
      responses:
        '201':
          description: The copy.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminLesson' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/lessons/{lessonId}/questions:
    parameters:
      - $ref: '#/components/parameters/LessonId'
    post:
      tags: [admin-lessons]
      summary: Create question
      operationId: adminCreateQuestion
      description: Creates the lesson quiz on first question; appended to the end.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminQuestionInput' }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminQuestion' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /admin/questions/{questionId}:
    parameters:
      - name: questionId
        in: path
        required: true
        schema: { type: integer }
    put:
      tags: [admin-lessons]
      summary: Update question
      operationId: adminUpdateQuestion
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminQuestionInput' }
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminQuestion' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    delete:
      tags: [admin-lessons]
      summary: Delete question
      operationId: adminDeleteQuestion
      responses:
        '204': { description: Deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/lessons/{lessonId}/tasks:
    parameters:
      - $ref: '#/components/parameters/LessonId'
    post:
      tags: [admin-lessons]
      summary: Create task
      operationId: adminCreateTask
      description: Appended to the end of the lesson's tasks.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminTaskInput' }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminTask' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /admin/tasks/{taskId}:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    put:
      tags: [admin-lessons]
      summary: Update task
      operationId: adminUpdateTask
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminTaskInput' }
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminTask' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    delete:
      tags: [admin-lessons]
      summary: Delete task
      operationId: adminDeleteTask
      description: Cascades to submissions.
      responses:
        '204': { description: Deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: session
      description: HttpOnly session cookie set by `POST /auth/login`.
  parameters:
...
    CourseId:
      name: courseId
      in: path
      required: true
      schema: { type: integer }
    LessonId:
      name: lessonId
      in: path
      required: true
      schema: { type: integer }
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema: { type: integer }
...
  requestBodies:
    Published:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [published]
            properties:
              published: { type: boolean }
    Move:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [direction]
            properties:
              direction: { type: string, enum: [up, down] }
...
  responses:
    Unauthorized:
      description: '`unauthorized` — no or expired session.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: '`forbidden` or `csrf_rejected`.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: '`not_found`.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationFailed:
      description: '`validation_failed` with per-field codes in `details.fields`.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    SlugTaken:
      description: '`slug_taken`.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Slug:
      type: string
      pattern: '^[a-z0-9][a-z0-9_-]*$'
      maxLength: 100
...
    Difficulty:
      type: string
      enum: [beginner, intermediate, advanced, expert]
    TaskDifficulty:
      type: string
      enum: [easy, medium, hard]
    LessonKind:
      type: string
      enum: [theory, quiz, lab, sim, sql]
    TaskKind:
      type: string
      enum: [shell, go]
    ContentFormat:
      type: string
      enum: [md, html]
...
    Error:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples: [unauthorized, forbidden, not_found, validation_failed, slug_taken, already_author, sandbox_disabled]
        message:
          type: string
          description: Developer-facing, not localized.
        details: { $ref: '#/components/schemas/ErrorDetails' }
    ErrorDetails:
      type: object
      properties:
        fields:
          type: object
          additionalProperties: { type: string }
          description: Field name → error code (`required`, `too_short`, `too_long`, `invalid_format`, `invalid_value`, `not_found`, `not_author`).
        issues:
          type: array
          items: { $ref: '#/components/schemas/ImportIssue' }
...
    GlossaryItem:
      type: object
      required: [term, definition]
      properties:
        term: { type: string }
        definition: { type: string }
    TestCase:
      type: object
      required: [input, expected_output]
      properties:
        input: { type: string }
        expected_output: { type: string }

    # ── simulators ──
...
    AdminLessonDetail:
      type: object
      required: [lesson, questions, tasks]
      properties:
        lesson: { $ref: '#/components/schemas/AdminLesson' }
        questions:
          type: array
          items: { $ref: '#/components/schemas/AdminQuestion' }
        tasks:
          type: array
          items: { $ref: '#/components/schemas/AdminTask' }
    AdminLessonInput:
      type: object
      required: [slug, title, kind, format, difficulty]
      properties:
        slug: { $ref: '#/components/schemas/Slug' }
        title: { type: string, minLength: 1, maxLength: 200 }
        kind: { $ref: '#/components/schemas/LessonKind' }
        format: { $ref: '#/components/schemas/ContentFormat' }
        difficulty: { $ref: '#/components/schemas/Difficulty' }
        content: { type: string, maxLength: 500000 }
        vm_image: { type: string, maxLength: 40 }
        vm_init: { type: string }
        published: { type: boolean, description: 'Omitted → unchanged (draft on create). Changing it needs `can_publish`.' }
    AdminLesson:
      type: object
      required: [id, course_id, slug, title, kind, format, difficulty, content, vm_image, vm_init, published, order_num, source, created_at]
      properties:
        id: { type: integer }
        course_id: { type: integer }
        slug: { type: string }
        title: { type: string }
        kind: { $ref: '#/components/schemas/LessonKind' }
        format: { $ref: '#/components/schemas/ContentFormat' }
        difficulty: { $ref: '#/components/schemas/Difficulty' }
        content: { type: string }
        vm_image: { type: string }
        vm_init: { type: string }
        published: { type: boolean }
        order_num: { type: integer }
        source: { type: string, enum: [seed, admin] }
        created_at: { type: string, format: date-time }
...
    AdminQuestionInput:
      type: object
      required: [question, options, correct_index]
      properties:
        question: { type: string, minLength: 1, maxLength: 5000, description: HTML allowed; sanitized on render. }
        options:
          type: array
          minItems: 2
          maxItems: 10
          items: { type: string, minLength: 1, maxLength: 2000 }
        option_explanations:
          type: array
          items: { type: string, maxLength: 5000 }
          description: Empty or the same length as `options`.
        correct_index: { type: integer, minimum: 0, description: 0-based; must be < options length. }
        explanation: { type: string, maxLength: 5000 }
    AdminQuestion:
      type: object
      required: [id, lesson_id, order_num, question, options, option_explanations, correct_index, explanation]
      properties:
        id: { type: integer }
        lesson_id: { type: integer }
        order_num: { type: integer }
        question: { type: string }
        options:
          type: array
          items: { type: string }
        option_explanations:
          type: array
          items: { type: string }
        correct_index: { type: integer }
        explanation: { type: string }
    AdminTaskInput:
      type: object
      required: [title, kind, format, difficulty]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        kind: { $ref: '#/components/schemas/TaskKind' }
        format: { $ref: '#/components/schemas/ContentFormat' }
        difficulty: { $ref: '#/components/schemas/TaskDifficulty' }
        description: { type: string, maxLength: 100000 }
        hints: { type: string, maxLength: 20000 }
        solution: { type: string, maxLength: 100000 }
        starter_code: { type: string, maxLength: 64000 }
        sandbox_image: { type: string, description: 'Shell tasks: one of the lab images, e.g. `golearn/sandbox:latest`.' }
        setup_script: { type: string, maxLength: 64000 }
        check_script: { type: string, maxLength: 64000 }
        glossary:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/GlossaryItem' }
        test_cases:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/TestCase' }
    AdminTask:
      type: object
      required: [id, lesson_id, order_num, title, kind, format, difficulty, description, hints, solution, starter_code, sandbox_image, setup_script, check_script, glossary, test_cases]
      properties:
        id: { type: integer }
        lesson_id: { type: integer }
        order_num: { type: integer }
        title: { type: string }
        kind: { $ref: '#/components/schemas/TaskKind' }
        format: { $ref: '#/components/schemas/ContentFormat' }
        difficulty: { $ref: '#/components/schemas/TaskDifficulty' }
        description: { type: string }
        hints: { type: string }
        solution: { type: string }
        starter_code: { type: string }
        sandbox_image: { type: string }
        setup_script: { type: string }
        check_script: { type: string }
        glossary:
          type: array
          items: { $ref: '#/components/schemas/GlossaryItem' }
        test_cases:
          type: array
          items: { $ref: '#/components/schemas/TestCase' }
...
    ImportIssue:
      type: object
      required: [level, code]
      properties:
        level: { type: string, enum: [error, warn] }
        code:
          type: string
          examples: [lesson_slug_empty, lesson_slug_duplicate, no_lessons, lesson_title_empty, unknown_kind, unknown_format, too_few_options, correct_out_of_range, check_without_image]
        lesson_slug: { type: string }
        path: { type: string, description: 'JSON pointer, e.g. `/lessons/2/quiz/0`.' }
...